Move acro and se logic into cogs
