--- defines/acro_choices.json
@@ -1,6 +1,0 @@
-{
-	"only pu words": "pu",
-	"pu and ku suli words": "ku suli",
-	"all pu, ku suli and ku lili words": "ku lili",
-	"all words": "all"
-}
