+++ src/ilo/cogs/acro/cog.py
@@ -2,12 +2,19 @@ from discord.ext import commands
 from discord.commands import slash_command
 from discord import Option
 
-from ilo.defines import acro_choices
 from ilo.defines import text
 from ilo.preferences import preferences
 from ilo.preferences import Template
 
 from ilo.cogs.acro import acronym
+
+
+acro_choices = {
+	"only pu words": "pu",
+	"pu and ku suli words": "ku suli",
+	"all pu, ku suli and ku lili words": "ku lili",
+	"all words": "all"
+}
 
 
 class CogAcro(commands.Cog):
