+++ src/ilo/jasima.py
@@ -1,10 +1,7 @@
 import json
 import urllib.request
 
-from ilo.defines import extraemoji
-
 JSON_LINK = "https://lipu-linku.github.io/jasima/data.json"
-LANGUAGE_OPTIONS_PATH = "slashcommands/language_options.json"
 
 help_message = "The word you requested, ***{}***, is not in the database I use. Make sure you didn't misspell it, or talk to kala Asi if this word really is missing."
 multiple_words_message = "The phrase you requested, ***{}***, contains multiple words. I am but a simple dictionary and can only do words one at a time."
@@ -25,17 +22,6 @@ def get_word_entry(word):
     if word not in entries:
         return help_message.format(word)
     return entries[word]
-
-
-def sitelen_emosi(word):
-    entries = bundle["data"]
-    if word in entries:
-        if "sitelen_emosi" in entries[word]:
-            return entries[word]["sitelen_emosi"]
-    chars = []
-    for letter in word:
-        chars.append(extraemoji[letter])
-    return " ".join(chars)
 
 
 def get_languages_for_slash_commands():
